+++ contracts/interfaces/ILiquidityCertificate.sol
@@ -17,46 +17,48 @@ interface ILiquidityCertificate {
     // get the metaDefender address
     function metaDefender() external view returns (address);
 
-    function totalValidCertificateLiquidity() external view returns (uint);
+    function totalValidCertificateLiquidity() external view returns (uint256);
 
-    function totalPendingCertificateLiquidity() external view returns (uint);
+    function totalPendingCertificateLiquidity() external view returns (uint256);
 
-    function MIN_LIQUIDITY() external view returns (uint);
+    function MIN_LIQUIDITY() external view returns (uint256);
 
     function getLiquidityProviders(
         address owner
-    ) external view returns (uint[] memory);
+    ) external view returns (uint256[] memory);
 
-    function getLiquidity(uint certificateId) external view returns (uint);
+    function getLiquidity(
+        uint256 certificateId
+    ) external view returns (uint256);
 
     function getCertificateInfo(
-        uint certificateId
+        uint256 certificateId
     ) external view returns (CertificateInfo memory);
 
     function updateRewardDebtEpochIndex(
-        uint certificateId,
+        uint256 certificateId,
         uint64 currentEpochIndex
     ) external;
 
-    function updateSPSLocked(uint certificateId, uint SPSLocked) external;
+    function updateSPSLocked(uint256 certificateId, uint256 SPSLocked) external;
 
     function newEpochCreated() external;
 
     function mint(
         uint64 enteredEpochIndex,
-        uint liquidity
-    ) external returns (uint);
+        uint256 liquidity
+    ) external returns (uint256);
 
-    function decreaseLiquidity(uint certificateId) external;
+    function decreaseLiquidity(uint256 certificateId) external;
 
-    function decreaseLiquidityByJudger(uint certificateId) external;
+    function decreaseLiquidityByJudger(uint256 certificateId) external;
 
-    function expire(uint certificateId, uint64 currentEpochIndex) external;
+    function expire(uint256 certificateId, uint64 currentEpochIndex) external;
 
     function expireByJudger(
-        uint certificateId,
+        uint256 certificateId,
         uint64 currentEpochIndex
     ) external;
 
-    function belongsTo(uint certificateId) external view returns (address);
+    function belongsTo(uint256 certificateId) external view returns (address);
 }
